Change zoom to increase and decrease image size by 5% rather than a static value, which stops the effect 'slowing down' and 'speeding up' as the image gets bigger and smaller respectively

+++ src/projection_control.c
@@ -20,7 +20,7 @@ static void	isometric_projection(t_draw *draw, int *x, int *y, int *z)
 	_y = (float) (*y) * draw->tile_width;
 	_z = (float) (*z) * draw->z_factor;	
 
-	diff = draw->tile_width - draw->orig_width;
+	diff = (float) draw->tile_width - draw->orig_width;
 	_z *= (diff / draw->orig_width) + 1;
 	*x = draw->x_offset + (int) ((_x - _y) * COS_30);
 	*y = draw->y_offset + (int) (-_z + (_x + _y) * SIN_30);
